Remove over-engineered abstraction of beatmap difficulty label
The future of how skinning should be implemented in song select is not quite determined yet, I don't think it's a good idea to overly abstract and split out every single minor component just for the sake that maybe it can become skinnable on its own. It's a better idea to implement things as-is and actually change how skinning works to work with that instead (or find some middle-point solution / third option etc.).

The existence of this component as a separate class is also questionable but I'll leave that for now.