Allow `BeatmapSetPanel` to determine expansion state
The `Carousel` is currently written in a manner agnostic from the concept of "group panels", so it's hard to add this state anywhere in the base code.

I've thought about potentially making `BeatmapCarousel` manage the expansion state of each set panel but that's also not quite possible due to the nature of the carousel code not allowing abstraction and derived classes (which I understand, to avoid too much OOP etc.).